feat: replace ids with bigint and bigserial

+++ src/main/resources/sql/authCreate.sql
@@ -1,23 +1,23 @@
 CREATE TABLE IF NOT EXISTS role
 (
-    role_id serial,
+    role_id bigserial,
     name    character varying(255) NOT NULL UNIQUE,
     PRIMARY KEY (role_id)
 );
 
 CREATE TABLE IF NOT EXISTS person_role
 (
-    person_id integer NOT NULL,
-    role_id   integer NOT NULL,
+    person_id bigint NOT NULL,
+    role_id   bigint NOT NULL,
     FOREIGN KEY (role_id) REFERENCES role (role_id) ON DELETE CASCADE,
     PRIMARY KEY (person_id, role_id)
 );
 
 CREATE TABLE IF NOT EXISTS refresh_token_info
 (
-    token_id        serial,
+    token_id        bigserial,
     refresh_token   character varying(255) NOT NULL UNIQUE,
-    person_id       integer                NOT NULL,
+    person_id       bigint                 NOT NULL,
     role            character varying(255) NOT NULL,
     expiration_time timestamp              NOT NULL,
     PRIMARY KEY (token_id)
