+++ src/main/resources/sql/feedbackCreate.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS keeper_feedback
 (
-    keeper_id   integer NOT NULL,
+    keeper_id   bigint  NOT NULL,
     explorer_id integer,
     rating      integer NOT NULL,
     comment     character varying(1000),
@@ -9,7 +9,7 @@ CREATE TABLE IF NOT EXISTS keeper_feedback
 CREATE TABLE IF NOT EXISTS explorer_feedback
 (
     explorer_id integer,
-    keeper_id   integer NOT NULL,
+    keeper_id   bigint  NOT NULL,
     rating      integer NOT NULL,
     comment     character varying(1000),
     PRIMARY KEY (explorer_id)
